Corrections to previous commit.

+++ src/game/ChatCommands/Level3.cpp
@@ -6294,7 +6294,7 @@ bool ChatHandler::HandleSendMassMailCommand(char* args)
     }
 
     // GM mail
-    MailSender sender(MAIL_NORMAL, 0, MAIL_STATIONERY_GM);
+    MailSender sender(MAIL_NORMAL, (uint32)0, MAIL_STATIONERY_GM);
 
     sMassMailMgr.AddMassMailTask(draft, sender, raceMask);
 
@@ -6398,7 +6398,7 @@ bool ChatHandler::HandleSendItemsCommand(char* args)
     if (!HandleSendItemsHelper(draft, args))
         { return false; }
 
-    MailSender sender(MAIL_NORMAL, 0, MAIL_STATIONERY_GM);
+    MailSender sender(MAIL_NORMAL, (uint32)0, MAIL_STATIONERY_GM);
 
     draft.SendMailTo(MailReceiver(receiver, receiver_guid), sender);
 
@@ -6428,7 +6428,7 @@ bool ChatHandler::HandleSendMassItemsCommand(char* args)
         return false;
     }
 
-    MailSender sender(MAIL_NORMAL, 0, MAIL_STATIONERY_GM);
+    MailSender sender(MAIL_NORMAL, (uint32)0, MAIL_STATIONERY_GM);
 
     sMassMailMgr.AddMassMailTask(draft, sender, raceMask);
 
@@ -6477,7 +6477,7 @@ bool ChatHandler::HandleSendMoneyCommand(char* args)
     if (!HandleSendMoneyHelper(draft, args))
         { return false; }
 
-    MailSender sender(MAIL_NORMAL, 0, MAIL_STATIONERY_GM);
+    MailSender sender(MAIL_NORMAL, (uint32)0, MAIL_STATIONERY_GM);
 
     draft.SendMailTo(MailReceiver(receiver, receiver_guid), sender);
 
@@ -6507,7 +6507,7 @@ bool ChatHandler::HandleSendMassMoneyCommand(char* args)
     }
 
     // from console show nonexistent sender
-    MailSender sender(MAIL_NORMAL, 0, MAIL_STATIONERY_GM);
+    MailSender sender(MAIL_NORMAL, (uint32)0, MAIL_STATIONERY_GM);
 
     sMassMailMgr.AddMassMailTask(draft, sender, raceMask);
